Only build from dump if we actually downloaded it (update).

+++ dump_update.sh
@@ -6,8 +6,10 @@ set -e
 
 # source dump_download.sh so that we set $TIMESTAMP
 source dump_download.sh
-time bash dump_build.sh $TIMESTAMP
+if $DOWNLOADED; then
+  time bash dump_build.sh $TIMESTAMP
 
-echo "Update default version"
-rm -f data/version/default
-ln -s $TIMESTAMP data/version/default
+  echo "Update default version"
+  rm -f data/version/default
+  ln -s $TIMESTAMP data/version/default
+fi
